Move some routing namespaces
