+++ src/Bundle/Resources/config/services/routing.xml
@@ -39,16 +39,16 @@
         </service>
         <service id="Sylius\Bundle\ResourceBundle\Routing\RoutesAttributesLoader" alias="sylius.routing.loader.routes_attributes" public="false" />
 
-        <service id="sylius.routing.factory.operation_route_name_factory" class="Sylius\Resource\Symfony\Routing\Factory\OperationRouteNameFactory" public="false" />
-        <service id="Sylius\Resource\Symfony\Routing\Factory\OperationRouteNameFactoryInterface" alias="sylius.routing.factory.operation_route_name_factory" />
+        <service id="sylius.routing.factory.operation_route_name_factory" class="Sylius\Resource\Symfony\Routing\Factory\RouteName\OperationRouteNameFactory" public="false" />
+        <service id="Sylius\Resource\Symfony\Routing\Factory\RouteName\OperationRouteNameFactoryInterface" alias="sylius.routing.factory.operation_route_name_factory" />
 
-        <service id="sylius.routing.factory.operation_route_path_factory.default" class="Sylius\Resource\Symfony\Routing\Factory\OperationRoutePathFactory" public="false" />
+        <service id="sylius.routing.factory.operation_route_path_factory.default" class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\OperationRoutePathFactory" public="false" />
 
         <service id="sylius.routing.factory.operation_route_path_factory" alias="sylius.routing.factory.operation_route_path_factory.default" />
-        <service id="Sylius\Resource\Symfony\Routing\Factory\OperationRoutePathFactoryInterface" alias="sylius.routing.factory.operation_route_path_factory" />
+        <service id="Sylius\Resource\Symfony\Routing\Factory\RoutePath\OperationRoutePathFactoryInterface" alias="sylius.routing.factory.operation_route_path_factory" />
 
         <service id="sylius.routing.factory.operation_route_path_factory.collection"
-                 class="Sylius\Resource\Symfony\Routing\Factory\CollectionOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\CollectionOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="60"
         >
@@ -56,7 +56,7 @@
         </service>
 
         <service id="sylius.routing.factory.operation_route_path_factory.create"
-                 class="Sylius\Resource\Symfony\Routing\Factory\CreateOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\CreateOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="-50"
         >
@@ -64,7 +64,7 @@
         </service>
 
         <service id="sylius.routing.factory.operation_route_path_factory.bulk_operation"
-                 class="Sylius\Resource\Symfony\Routing\Factory\BulkOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\BulkOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="-40"
         >
@@ -72,7 +72,7 @@
         </service>
 
         <service id="sylius.routing.factory.operation_route_path_factory.update"
-                 class="Sylius\Resource\Symfony\Routing\Factory\UpdateOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\UpdateOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="-30"
         >
@@ -80,7 +80,7 @@
         </service>
 
         <service id="sylius.routing.factory.operation_route_path_factory.delete"
-                 class="Sylius\Resource\Symfony\Routing\Factory\DeleteOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\DeleteOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="-20"
         >
@@ -88,7 +88,7 @@
         </service>
 
         <service id="sylius.routing.factory.operation_route_path_factory.show"
-                 class="Sylius\Resource\Symfony\Routing\Factory\ShowOperationRoutePathFactory"
+                 class="Sylius\Resource\Symfony\Routing\Factory\RoutePath\ShowOperationRoutePathFactory"
                  decorates="sylius.routing.factory.operation_route_path_factory.default"
                  decoration-priority="-10"
         >
