+++ src/Component/spec/Metadata/Resource/Factory/RedirectResourceMetadataCollectionFactorySpec.php
@@ -24,7 +24,7 @@ use Sylius\Resource\Metadata\Resource\ResourceMetadataCollection;
 use Sylius\Resource\Metadata\ResourceMetadata;
 use Sylius\Resource\Metadata\Show;
 use Sylius\Resource\Metadata\Update;
-use Sylius\Resource\Symfony\Routing\Factory\OperationRouteNameFactory;
+use Sylius\Resource\Symfony\Routing\Factory\RouteName\OperationRouteNameFactory;
 
 final class RedirectResourceMetadataCollectionFactorySpec extends ObjectBehavior
 {
