+++ src/Component/spec/Symfony/Routing/Factory/RoutePath/CollectionOperationRoutePathFactorySpec.php
@@ -11,13 +11,13 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Resource\Symfony\Routing\Factory;
+namespace spec\Sylius\Resource\Symfony\Routing\Factory\RoutePath;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Resource\Metadata\Api;
 use Sylius\Resource\Metadata\Index;
-use Sylius\Resource\Symfony\Routing\Factory\CollectionOperationRoutePathFactory;
-use Sylius\Resource\Symfony\Routing\Factory\OperationRoutePathFactoryInterface;
+use Sylius\Resource\Symfony\Routing\Factory\RoutePath\CollectionOperationRoutePathFactory;
+use Sylius\Resource\Symfony\Routing\Factory\RoutePath\OperationRoutePathFactoryInterface;
 
 final class CollectionOperationRoutePathFactorySpec extends ObjectBehavior
 {
