+++ src/Component/src/Symfony/Routing/Factory/RouteName/OperationRouteNameFactoryInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Resource\Symfony\Routing\Factory;
+namespace Sylius\Resource\Symfony\Routing\Factory\RouteName;
 
 use Sylius\Resource\Metadata\Operation;
