+++ src/Component/src/Symfony/Routing/Factory/RoutePath/DeleteOperationRoutePathFactory.php
@@ -11,12 +11,13 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Resource\Symfony\Routing\Factory;
+namespace Sylius\Resource\Symfony\Routing\Factory\RoutePath;
 
-use Sylius\Resource\Metadata\CreateOperationInterface;
+use Sylius\Resource\Metadata\Api\ApiOperationInterface;
+use Sylius\Resource\Metadata\DeleteOperationInterface;
 use Sylius\Resource\Metadata\HttpOperation;
 
-final class CreateOperationRoutePathFactory implements OperationRoutePathFactoryInterface
+final class DeleteOperationRoutePathFactory implements OperationRoutePathFactoryInterface
 {
     public function __construct(private OperationRoutePathFactoryInterface $decorated)
     {
@@ -25,15 +26,12 @@ final class CreateOperationRoutePathFactory implements OperationRoutePathFactory
     public function createRoutePath(HttpOperation $operation, string $rootPath): string
     {
         $shortName = $operation->getShortName();
+        $identifier = $operation->getResource()?->getIdentifier() ?? 'id';
 
-        if ($operation instanceof CreateOperationInterface) {
-            $path = match ($shortName) {
-                'create' => '/new',
-                'post' => '',
-                default => '/' . $shortName,
-            };
+        if ($operation instanceof DeleteOperationInterface) {
+            $path = $operation instanceof ApiOperationInterface && 'delete' === $shortName ? '' : '/' . $shortName;
 
-            return sprintf('%s%s', $rootPath, $path);
+            return sprintf('%s/{%s}%s', $rootPath, $identifier, $path);
         }
 
         return $this->decorated->createRoutePath($operation, $rootPath);
