+++ src/Component/src/Symfony/Routing/Factory/RoutePath/ShowOperationRoutePathFactory.php
@@ -11,13 +11,12 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Resource\Symfony\Routing\Factory;
+namespace Sylius\Resource\Symfony\Routing\Factory\RoutePath;
 
-use Sylius\Resource\Metadata\Api\ApiOperationInterface;
-use Sylius\Resource\Metadata\DeleteOperationInterface;
 use Sylius\Resource\Metadata\HttpOperation;
+use Sylius\Resource\Metadata\ShowOperationInterface;
 
-final class DeleteOperationRoutePathFactory implements OperationRoutePathFactoryInterface
+final class ShowOperationRoutePathFactory implements OperationRoutePathFactoryInterface
 {
     public function __construct(private OperationRoutePathFactoryInterface $decorated)
     {
@@ -28,8 +27,11 @@ final class DeleteOperationRoutePathFactory implements OperationRoutePathFactory
         $shortName = $operation->getShortName();
         $identifier = $operation->getResource()?->getIdentifier() ?? 'id';
 
-        if ($operation instanceof DeleteOperationInterface) {
-            $path = $operation instanceof ApiOperationInterface && 'delete' === $shortName ? '' : '/' . $shortName;
+        if ($operation instanceof ShowOperationInterface) {
+            $path = match ($shortName) {
+                'show', 'get' => '',
+                default => '/' . $shortName,
+            };
 
             return sprintf('%s/{%s}%s', $rootPath, $identifier, $path);
         }
